Add nix flake support for the repo's quake tools (tools like q3map)
A nix flake allows building and running the software with the Nix
package manager, used in NixOS or standalone. You can consider as
a package or repo description.

In this case, this only packages the quake tool binaries, that is
h2data, q2map, q3data, q3map2 and qdata3.