Make it build on M1 Mac
The CMake changes are to work around the issue that the pkgconfig
module is populating the xxx_LIBRARIES variable with the library
name only (e.g. 'glib-2.0') rather than the full path, which the
linker is unable to find. I guess the correct way to do it is to
use the xxx_LDFLAGS variable from pkgconfig which should tell the
linker where to find the libraries, but it's kind of annoying
because we also have a non-pkgconfig route.

The xxx_LINK_LIBRARIES variable has the full library paths but
it is available only from CMake 3.12 so it can't be used
unconditionally.