plugins: chose Radiant IQM plugin, iqmmodel for macOS, picomodel for everything else
The picomodel IQM code makes NetRadiant crash on macOS.
No one has found time to investigate this in month.

We need a working NetRadiant IQM plugin to load
Unvanquished game entity models on macOS, the
plugin from AAradiant (AlienArena project) is
known to work.

Q3map2 uses picomodel to load IQM on all platforms
because iqmmodel plugin is radiant-only, so there
may be a remaining bug when baking IQM models in BSP
with q3map2 on macOS (not tested), but editing a map
with IQM entity models and misc_animated_models will
not crash NetRadiant.

This patch makes possible to release a macOS NetRadiant
build that don't crash when rendering IQM models.

We should fix picomodel crash with IQM on macOS in any way,
because it is believed q3map2 may need it, but waiting for
a picomodel fix will not delay any NetRadiant release.

One can force NetRadiant to be built against a given IQM
plugin (for example to debug it) whatever the platform
by setting the RADIANT_IQM_PLUGIN CMake variable to
either `iqmmodel` or `picomodel`. Any other value,
including empty string, will set the default plugin for
the platform.