Reduce patch drawsurf references
If patch is definetly on one side of plane, then it is not on the other
unless proven otherwise. This theoretically can mess-up with patch
collision in other(than DarkPlaces) engines that use it for collision(no idea how)
if part of patch is barely inside of visleaf.

New behaviour needs more testing and maybe put behind flag/argument.